Move environment setup and breakdown to a separate class.
