+++ .gitignore
@@ -1,2 +1,3 @@
 ﻿# Default ignored files
 .idea
+Notes about docker.txt
